+++ src/modules/shared/services/api/index.ts
@@ -1,2 +1,2 @@
 export * from './interfaces';
-export * from './Axios';
+export * from './Http';
